Get abstract sockets working
- Remove support for using abstract sockets on hosts, as we only
  want to support that on Android.
- Remove #ifdefs and split the code up to target specific files
