--- renderdoc/core/net_cfg.h
@@ -1,13 +1,0 @@
-
-#pragma once
-
-/* Abstract socket config */
-// Enable defines to allow abstract socket servers on the given platform
-#define LINUX_ABSTRACT_SOCKET
-#define ANDROID_ABSTRACT_SOCKET
-
-// Enable defines to allow abstract socket client connection on the given platform
-#define LINUX_ABSTRACT_SOCKET_CLIENT
-#define ANDROID_ABSTRACT_SOCKET_CLIENT
-
-
