+++ renderdoc/os/posix/linux/linux_process.cpp
@@ -24,7 +24,6 @@
 
 #include <unistd.h>
 #include "os/os_specific.h"
-#include "core/net_cfg.h"
 
 extern char **environ;
 
@@ -38,17 +37,9 @@ char **GetCurrentEnvironment()
 
 int GetIdentPort(pid_t childPid)
 {
-  const char *net_target =
-#ifdef LINUX_ABSTRACT_SOCKET
-    "unix"
-#else
-    "tcp"
-#endif
-    ;
-
   int ret = 0;
 
-  string procfile = StringFormat::Fmt("/proc/%d/net/%s", (int)childPid, net_target);
+  string procfile = StringFormat::Fmt("/proc/%d/net/tcp", (int)childPid);
 
   int waitTime = INITIAL_WAIT_TIME;
 
@@ -76,28 +67,9 @@ int GetIdentPort(pid_t childPid)
       line[sz - 1] = 0;
       fgets(line, sz - 1, f);
 
-#ifdef LINUX_ABSTRACT_SOCKET
-      int port = 0;
-      char* startpos = strstr(line, "@/renderdoc/");
-
-      if (startpos == NULL)
-      {
-        // This is not the line you are looking for.
-        continue;
-      }
-
-      int num = sscanf(startpos, "@/renderdoc/%d", &port);
-
-      // find open abstract listen socket on path: /renderdoc/<port>
-      if(num == 1 && port >= RenderDoc_FirstTargetControlPort &&
-         port <= RenderDoc_LastTargetControlPort)
-      {
-        ret = port;
-        break;
-      }
-#else
       int socketnum = 0, hexip = 0, hexport = 0;
       int num = sscanf(line, " %d: %x:%x", &socketnum, &hexip, &hexport);
+
       // find open listen socket on 0.0.0.0:port
       if(num == 3 && hexip == 0 && hexport >= RenderDoc_FirstTargetControlPort &&
          hexport <= RenderDoc_LastTargetControlPort)
@@ -105,7 +77,6 @@ int GetIdentPort(pid_t childPid)
         ret = hexport;
         break;
       }
-#endif
     }
 
     FileIO::fclose(f);
